- basic SM4 implementation - common interface for SM2 and SM4

+++ sralgorithms/src/com/ideabulbs/sralgorithms/sm/test/TestSM2.java
@@ -14,7 +14,7 @@ public class TestSM2 {
 	public void testIntervalOne() {
 		SM2 sm2 = new SM2(2.5f, 4);
 		int newInterval = sm2.getNextInterval(1);
-		float newEF = sm2.getNewEFactor();
+		double newEF = sm2.getNewEFactor();
 		assertEquals(newInterval, 1);
 		assertTrue((2.4 < newEF) && (newEF < 2.6));
 	}
@@ -37,15 +37,15 @@ public class TestSM2 {
 		grades.add(4);
 		grades.add(5);
 
-		Vector<Float> efs = new Vector<Float>();
-		efs.add(2.5f);
+		Vector<Double> efs = new Vector<Double>();
+		efs.add(2.5);
 
 		Vector<Integer> intervals = new Vector<Integer>();
 		intervals.add(1);
 		
 		for (int i=0; i<grades.size(); i++) {
 			SM2 sm2 = new SM2(efs.get(i), grades.get(i));
-			float newEF = sm2.getNewEFactor();
+			double newEF = sm2.getNewEFactor();
 			int newInterval = sm2.getNextInterval(i);
 			efs.add(newEF);
 			intervals.add(newInterval);
@@ -68,7 +68,7 @@ public class TestSM2 {
 		assertTrue(intervals.get(14) == 17);
 		
 		System.out.print("EF : ");
-		for (Float f : efs) {
+		for (Double f : efs) {
 			System.out.print(String.format("%.2f", f) + "\t");
 		}
 		System.out.println();
